requirements.txt
beautifulsoup4==4.7.1
Click==7.0
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.0
MarkupSafe==1.1.0
pbr==5.1.2
pyaml==18.11.0
pylinkvalidator==0.3
PyYAML==3.13
six==1.12.0
smmap2==2.0.5
soupsieve==1.7.3
stevedore==1.30.0
Werkzeug==0.14.1
